PAA-13 : fixing global state
fixing user and address global state
